requirements.txt
flask==1.0.2
flask-sqlalchemy
scrapy==1.5.1